refactor: split conform.rs and main.rs into focused submodules
- crates/mogen-dsl/src/conform.rs (1310) → conform/{mod,spec,kinds,path,
  patch,place,tests}.rs; mod.rs holds the public entry points and the
  apply_conform dispatcher, with each mode kept on its own.
- crates/mogen/src/main.rs (1278) → cli/{value_args,auth,moghub,cmd}.rs
  + a slim main.rs (288 lines) that's just `Cli` and the dispatch
  match. Largest leaf is cmd.rs (the big Cmd enum with all clap flags).

No behavioural changes. Verified:
- mogen-dsl test suite (215 tests, including the 20 conform tests) all pass.
- mogen --help / auth --help / generate --help render the same surface.
